refactor: split views/metadata.py by role
The metadata module served both roles in one file. Split it:

- authorization_server/views/metadata.py: RFC 8414 OAuthServerMetadataView plus
  the shared discovery helpers (ServerMetadataViewMixin, bcp_filter_*), also used
  by OIDC discovery.
- resource_server/views/metadata.py: RFC 9728 OAuthProtectedResourceMetadataView.

oauth2_provider/views/metadata.py becomes a compat shim re-exporting both (with
an __all__ and a DeprecationWarning). Internal importers (views/oidc.py,
views/__init__.py, the role facades) are repointed to the canonical modules.

Also add __all__ to the views/mixins.py shim so its intentional re-exports are
not reported as unused imports.

Full non-e2e suite passes (1033); role facades emit no first-party warnings.
